+++ elasticache/subnet-group.tf
@@ -1,4 +1,4 @@
 resource "aws_elasticache_subnet_group" "main" {
-  name       = "${var.project}-${var.environment}-elasticache"
+  name       = local.elasticache_name
   subnet_ids = var.subnet_ids
 }