+++ stack/app/elasticache.tf
@@ -12,6 +12,7 @@ module "elasticache" {
   # optional
   node_type                = var.elasticache_node_type
   snapshot_retention_limit = var.elasticache_snapshot_retention_limit
+  data_tiering_enabled     = var.elasticache_data_tiering_enabled
 
   replicas_per_node_group = var.elasticache_replicas_per_node_group
   shard_count             = var.elasticache_shards_per_replication_group