+++ stack/setup/variables.tf
@@ -15,6 +15,11 @@ variable "add_wildcard_subdomains" {
   default = true
 }
 
+variable "is_read_replica_on_same_domain" {
+  type    = bool
+  default = false
+}
+
 variable "eips_nat_count" {
   type    = number
   default = 1
@@ -28,4 +33,11 @@ variable "alternative_domains" {
 variable "kms_deletion_window_in_days" {
   type    = number
   default = 30
+}
+
+
+# in case this stack has a read-replica RDS, we need to access the master DB KMS key
+variable "rds_cross_region_kms_key_arn" {
+  type    = string
+  default = null
 }