+++ CHANGELOG.md
@@ -8,8 +8,28 @@ This project does not follow SemVer, since modules are independent of each other
 
 ## [unreleased]
 
+## ElastiCache
+- added `data_tiering_enabled` to the available options, defaulting to `false`
+- added `name` as optional parameter in case multiple Redis-Clusters need to be launched for the same project/environment
+
+## [v2022.07.08]
+
+## VPN
+- `availability_zones` can now be explicitly passed into the module, [#98](https://github.com/dbl-works/terraform/pull/98)
+
+## KMS Key Replica
+- new module that allows to create a replica of a KMS key in another region, [#97](https://github.com/dbl-works/
+
+## RDS
+- added the option to create a read-replica, [#97](https://github.com/dbl-works/terraform/pull/97)
+- allow to create unique names across regions by setting `regional`, [#97](https://github.com/dbl-works/terraform/pull/97)
+
+## VPC-Peering
+- added a new module `vpc-peering` to create a VPC Peering Resource, [#97](https://github.com/dbl-works/terraform/pull/97)
+
 ## Stack & Cloudflare
 - fixed issues when setting up a bastion, [#93](https://github.com/dbl-works/terraform/pull/93)
+- added the option to launch a stack with a RDS read-replica, [#97](https://github.com/dbl-works/terraform/pull/97)
 
 
 ## [v2022.06.23]